+++ derive_macros/src/stream.rs
@@ -1,0 +1,5 @@
+mod stream_newtype;
+mod stream_record_tuple;
+mod stream_tag;
+
+pub(crate) use self::{stream_newtype::*, stream_record_tuple::*, stream_tag::*};
